refactor: `ApiResponse`와 `ApiError`를 `useStuQuery`로 이동
