+++ src/pages/human-rights/schema.ts
@@ -65,35 +65,6 @@ export type HumanRightsComment = z.output<typeof HumanRightsCommentSchema>;
  */
 export type PostAcl = z.infer<typeof PostAclSchema>;
 
-// TODO: Move ApiResponse and ApiError to global scope
-/**
- * API의 기본 반환 응답입니다.
- * @typeParam T - 요청이 성공하였을 때 반환할 데이터
- */
-export interface ApiResponse<T> {
-  code: string;
-  message: string;
-  data: T;
-  isSuccess: boolean;
-}
-
-/**
- * API 오류 시 응답입니다.
- */
-export interface ApiError {
-  code: string;
-  message: string;
-  data: null;
-  isSuccess: false;
-}
-
-/**
- * 인권신고게시판의 단건 조회 데이터입니다.
- */
-export interface HumanRightsBoardPostResponse {
-  postDetailResDto: HumanRightsPostResponse;
-}
-
 /**
  * 인권신고게시판의 게시글 댓글 목록 데이터입니다.
  */
